Make iron ore a little harder to find, and in slightly smaller quantities
During play testing, it was found that the prior settings for iron ore
generation made finding large amounts of iron too easy. This is a bit of a nerf
in 2 ways, there is about 5% less iron in the world, and it's in more, slightly
smaller chunks. The spawn levels was also increased from a 24 block height to
32 block. It may be changed again in the future, depending on the results of
actual survival world play testing.