New Ore generation iteration 2
Here is the second iteration of the new ore generation, basically I follow a
lot of the same lines as my previous attempt, but I am disabling all of the TE
ores and vanilla ores and add them as custom generation types. I do this in
order to use the additional generator that is offered by CoFHCore, and this is
the NORMAL distribution method. This basically will make the ores, rather than
just randomly picking a place to generate between the minimum and maximum
values, it is more weighted to be, on average, in the center of the
distribution, the so called MeanY. This also offers a Y variance, which is the
number of levels above and below the MeanY that the ores can generate.

This generator is not without it's quirks, I have noticed through testing that
it is more skewed towards the bottom end of the spectrum, rather than being
truly in the center, and that it has more distribution on the bottom rather
than at the top. But this still does offer a better way to make ores appear in
more distinct bands, which is what I was attempting to accomplish with this ore
generation. The only thing to keep in mind is that it's usually skewed 2-3
blocks below what the MeanY variable is set at, and usually has slightly more
on the bottom end than at the top end.

How this Ore is laid out is as follows:
Copper
Iron
Tin
Lead
Silver
Gold
Ferrous/Redstone
Diamond is unaffected
Osmium is centered between gold and redstone, but can appear up into the
silver/lead layers and down below near the bottom of the world.

The actual arrangement of the layers isn't all that different from how it
normally is, but there is more ores in distinct bands, and there is far less
overlap between the ores than there was previously.