Change big reactors world gen and some additional balancing
Changed the world generation of big reactors, Yellorite ore, to spawn in the 9
layers above bedrock, so in order to mine this ore you'll need to perform
mining near bedrock, dealing with whatever lava hazards there exist. Also big
reactors now take steel for all recipes that require iron, the only difference
between this and the tppitweaks one is that steel is required for the fuel rods
as well. To help counteract this, reactor output has been bumped by 2.25 times,
to help offset the costs these changes make.