Change AE Ore spawn
AE certus ore spawning has been changed to be managed by CoFHCore, this value
is close to the FTB new defaults that they use, which is roughly 20-30% or so
more than default AE. AE ores now spawn in somewhat infrequent, larger chunks
than normal between the levels of 36 and 54. This is a uniform distribution, so
this means in play that it will be interspersed with the iron and tin layers,
with some sometimes showing up near the bottom of the copper layers.