Reenable Forge Hammer but disable hammering of higher tier ingots
The default gregtech hammer doubles the ingot requirements for plates and
returns only half of the normal item casings, which for some users may make
things too costly until they get a plate bending machine. Forge Hammers have
been reenabled but the manual hammering of all materials but copper, tin, iron,
and bronze has been disabled to enable needing to use a metal forming machine
or plate bender to get the higher tier machines. The only affect of this on
vanilla IC2 is the gold cables, but by the time you are making an MFE or above,
you should already be using a machine to make your plates. This also will
reduce the use of AE autocrafting with the forge hammer to have faster crafting
speeds for higher tier items.