Match values for dim trans with TE3 tesseract
The TE3 tesseract currently is made to be completely lossless (for energy, liquids, and items) and currently there is no plans from TeamCoFH to include config options that allow changing of the values. This change will make the dim transceiver match the TE3 tesseract values of 0 since there is no incentive to use it because the recipe is set to hard and it has loss of energy and uses energy to keep it's connection active.