Rebalance Magical Crops
Magical crops is being rebalanced in a couple of ways, first is the fact
that for most seed types, it will take 4x the amount of seeds to get the
same output that it used to, this is to slightly nerf automated mining,
as more expensive setups will be needed, and to be able to have large
fields growing resources. This obviously necessitates the need for more
seeds to be gotten per craft, and also that it will roughly be 4x less
output gotten per seed type. There will be no second seed drop chance,
so you'll need to invest in essence crops to get a renewable source of
essence to grow your fields. MFR support is also enabled, but is
slightly more balanced in the fact that it will not pick up any of the
extra drops that crops can drop, such as nature essence or essence
dusts.